refactor: 게임 모드 선택 컴포넌트 상태
- 상태를 외부에서 받도록 변경
- 새로고침 방지 hook으로 분리
