+++ packages/frontend/src/hooks/usePreventClose.tsx
@@ -1,0 +1,16 @@
+import { useEffect } from "react";
+
+const usePreventClose = () => {
+  useEffect(() => {
+    const preventClose = (e: BeforeUnloadEvent) => {
+      e.preventDefault();
+      e.returnValue = "";
+    };
+    window.addEventListener("beforeunload", preventClose);
+    return () => {
+      window.removeEventListener("beforeunload", preventClose);
+    };
+  }, []);
+};
+
+export default usePreventClose;
