+++ packages/frontend/src/pages/RoomPage/RoomPage.component.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { Navigate, useNavigate } from "react-router";
 
 import { useAtom, useAtomValue, useSetAtom } from "jotai";
@@ -15,6 +15,8 @@ import { Button } from "../../components/common/Button";
 import GameModeSegmentedControl from "../../components/GameModeSegmentedControl/GameModeSegmentedControl.component";
 import { Logo } from "../../components/Logo/Logo.component";
 import PlayerList from "../../components/PlayerList/PlayerList.component";
+import { GameMode, GAME_MODE_LIST } from "../../constants/game-mode";
+import usePreventClose from "../../hooks/usePreventClose";
 import { gameInfoAtom } from "../../store/game";
 import { playersAtom } from "../../store/players";
 import { roomIdAtom } from "../../store/roomId";
@@ -28,6 +30,8 @@ const RoomPage = () => {
   const [players, setPlayers] = useAtom(playersAtom);
   const setGameInfo = useSetAtom(gameInfoAtom);
   const navigate = useNavigate();
+  const [gameMode, setGameMode] = useState<GameMode>(GAME_MODE_LIST[0]);
+  usePreventClose();
 
   const onInviteClick = () => {
     const inviteUrl = `${window.location.origin}/?invite=${roomId}`;
@@ -39,19 +43,8 @@ const RoomPage = () => {
 
   const onGameStartClick = () => {
     if (!isHost) return;
-    socket.emit("game-start", { gameMode: "catchMind" });
+    socket.emit("game-start", { gameMode });
   };
-
-  useEffect(() => {
-    const preventClose = (e: BeforeUnloadEvent) => {
-      e.preventDefault();
-      e.returnValue = "";
-    };
-    window.addEventListener("beforeunload", preventClose);
-    return () => {
-      window.removeEventListener("beforeunload", preventClose);
-    };
-  }, []);
 
   useEffect(() => {
     const gameStartListener = (gameStartResponse: GameInfo) => {
@@ -82,7 +75,10 @@ const RoomPage = () => {
       <RoomPageContentBox>
         <PlayerList maxPlayer={10} players={players} sizeType="large" />
         <RoomPageRightContentBox>
-          <GameModeSegmentedControl />
+          <GameModeSegmentedControl
+            selectedGameMode={gameMode}
+            setSelectedGameMode={setGameMode}
+          />
           <RoomPageButtonBox>
             <Button variant="medium" onClick={onInviteClick}>
               초대
